+++ frontend/src/page/owner/index/OwnerIndex.css
@@ -1,0 +1,8 @@
+.owners-main {
+    display: flex;
+}
+
+.content {
+    flex-grow: 1;
+    display: flex;
+}
